# Changed All most all UI components # Added Filters to all # Completed Student Side # Completed Sidebar for mobile view # Tuned some components for Mobile view # Added profile picture upload feature

+++ student_lor/api.py
@@ -1,6 +1,7 @@
 # Create your views here.
 import json
 import os
+from datetime import timezone
 
 from django.core.exceptions import ObjectDoesNotExist
 from django.core.mail import send_mail
@@ -14,7 +15,7 @@ from rest_framework.views import APIView
 from tracker_final.settings import BASE_DIR
 from .permissions import HasGroupPermission
 from .serializers import *
-from .validations.validate_lor_submission import validate_lor_submission
+from .validations.validate_lor_submission import validate_lor_submission, validate_edit_lor_submission
 
 WITHDRAW_TEMPLATE = """
 Dear {{faculty.first_name}} {{faculty.last_name}}:
@@ -213,7 +214,9 @@ class EditLor(generics.GenericAPIView):
 			applications = FacultyListLor.objects.filter(lor=lor_details.id)
 			for item in applications:
 				template = Template(EDITLOR_TEMPLATE)
-				faculty_details = AppUser.objects.get(id=item.faculty)
+				print(item.faculty.id)
+				faculty_details = GetFacultyListSerializer(AppUser.objects.get(id=item.faculty.id)).data
+				print(faculty_details)
 				details = StudentDetails.objects.get(user=self.request.user.id)
 				send_mail(
 					'Lor Updated: Student has changed the details of LOR ',
@@ -221,7 +224,7 @@ class EditLor(generics.GenericAPIView):
 						context=Context(
 							{'faculty': faculty_details, 'student': details, 'deadline': lor_details.deadline})),
 					'ghotden@gmail.com',
-					[faculty_details.email],
+					[faculty_details["email"]],
 					fail_silently=False,
 				)
 			return Response({
@@ -294,6 +297,60 @@ class AddFacultyForLOR(generics.GenericAPIView):
 		return Response({"success": True})
 
 
+class EditSubmittedLorCourseDetails(generics.GenericAPIView):
+	permission_classes = [
+		permissions.IsAuthenticated,
+		HasGroupPermission
+	]
+	required_groups = {
+		'POST': ['student'],
+	}
+	serializer_class = EditSubmittedLorDetailsSerializer
+
+	def post(self, request, faculty, lor):
+		serializer = self.get_serializer(data=request.data)
+		print('received:', request.data)
+		validate_edit_lor_submission(data=request.data)
+		selected_lor = Lor.objects.get(id=lor)
+		if selected_lor.deadline.replace(tzinfo=None) <= datetime.now().replace(tzinfo=None):
+			errors = {
+				"lor": 'Your LOR is expired, choose another one'}
+			print(errors)
+			raise ValidationError(errors)
+		else:
+			check_exists = FacultyListLor.objects.filter(faculty=faculty, lor=lor)
+			if check_exists.count() > 0:
+				print(';;;;')
+				data = request.data
+				updated_details = FacultyListLor.objects.filter(lor_id=lor,
+																faculty_id=faculty).update(courses_done=data['courses_done'],
+					projects_done=data['projects_done'],
+					thesis_done=data['thesis_done'],
+					status=data['status'],
+					others=data['others'])
+				template = Template(EDITLOR_TEMPLATE)
+				lor = Lor.objects.get(id=lor)
+				dead_line = lor.deadline.replace(tzinfo=timezone.utc).astimezone(tz=None)
+				faculty_details = AppUser.objects.get(id=faculty)
+				details = StudentDetails.objects.get(user=self.request.user.id)
+				print('here')
+				send_mail(
+					'Lor Updated: Student has updated the Courses, Project details of Request',
+					template.render(
+						context=Context(
+							{'faculty': faculty_details, 'student': details, 'deadline': dead_line})),
+					'ghotden@gmail.com',
+					[faculty_details.email],
+					fail_silently=False,
+				)
+			else:
+				creation_errors = {
+					'application': 'Lor already requested'
+				}
+				ValidationError(creation_errors)
+		return Response({"success": True})
+
+
 # View Saved Lor
 class GetMySavedLor(APIView):
 	permission_classes = [
@@ -306,6 +363,11 @@ class GetMySavedLor(APIView):
 	serializer_class = ViewSavedLor
 
 	def get(self, request):
+		update_entries = Lor.objects.filter(user=self.request.user.id, deadline__lte=datetime.now(timezone.utc)
+											, expired=False)
+		for entry in update_entries:
+			new_entries = FacultyListLor.objects.filter(lor=entry["id"],
+														   application_status__in=['AC', 'AP']).update(application_status='EX')
 		update_entries = Lor.objects.filter(user=self.request.user.id, deadline__lte=datetime.now(timezone.utc)
 											, expired=False).update(expired=True)
 		entries = Lor.objects.filter(user=self.request.user.id, hidden=False)
@@ -359,20 +421,28 @@ class GetAppliedLor(APIView):
 	serializer_class = ViewAppliedFacultyListSerializer
 
 	def get(self, request):
+		update_entries = Lor.objects.filter(user=self.request.user.id, deadline__lte=datetime.now(timezone.utc), expired=False)
+		# print(update_entries)
+		for entry in update_entries:
+			new_entries = FacultyListLor.objects.filter(lor_id=entry.id,
+														   application_status__in=['AC', 'AP']).update(
+				application_status='EX')
+		update_entries = Lor.objects.filter(user=self.request.user.id, deadline__lte=datetime.now(timezone.utc)
+											, expired=False).update(expired=True)
 		entries = Lor.objects.filter(user=self.request.user.id)
 		result = []
 		for entry in entries:
 			str_data = serialize('json', FacultyListLor.objects.filter(lor=entry.id))
 			data = json.loads(str_data)
 			for item in data:
-				item["fields"]["lor"] = json.loads(serialize('json', Lor.objects.filter(id=item["fields"]["lor"])))[0][
-					"fields"]
+				print(json.loads(serialize('json', Lor.objects.filter(id=item["fields"]["lor"])))[0][
+						  "fields"])
+				item["fields"]["lor"] = ViewSavedLor(Lor.objects.get(id=item["fields"]["lor"])).data
 				item["fields"]["faculty"] = AppUser.objects.values("id", "email", "first_name", "last_name",
 																   "department_name").filter(
 					id=item["fields"]["faculty"])[0]
 				result.append(item["fields"])
 
-		print(result)
 		return Response(result)
 
 
@@ -390,50 +460,19 @@ class WithdrawApplications(APIView):
 		template = Template(WITHDRAW_TEMPLATE)
 		faculty_details = AppUser.objects.get(id=faculty)
 		details = StudentDetails.objects.get(user=self.request.user.id)
-		FacultyListLor.objects.get(lor=lor, faculty=faculty).delete()
-		send_mail(
-			'Request Withdrawn: Lor Request has been Withdrawn by the student',
-			template.render(context=Context({'faculty': faculty_details, 'student': details})),
-			'ghotden@gmail.com',
-			[faculty_details.email],
-			fail_silently=False,
-		)
-		return Response('Application Withdrawn')
-
-
-class EditLorApplication(generics.GenericAPIView):
-	permission_classes = [
-		permissions.IsAuthenticated,
-		HasGroupPermission
-	]
-	required_groups = {
-		'POST': ['student'],
-	}
-	serializer_class = CreateLorRequestSerializer
-
-	def post(self, request, lor):
-		try:
-			existing = Lor.objects.get(id=lor, user=self.request.user.id)
-			serializer = self.get_serializer(existing, data=request.data)
-			serializer.is_valid(raise_exception=True)
-			lor = FacultyListLor.objects.filter(lor=lor)
-			student_details = serializer.save()
-			for item in lor:
-				faculty_details = AppUser.objects.get(id=item.faculty)
-				details = StudentDetails.objects.get(user=self.request.user.id)
-				dead_line = lor.deadline.replace(tzinfo=timezone.utc).astimezone(tz=None)
-				template = Template(EDITLOR_TEMPLATE)
-				send_mail(
-					'Lor Application Edited by the student',
-					template.render(
-						context=Context({'faculty': faculty_details, 'student': details, 'deadline': dead_line})),
-					'ghotden@gmail.com',
-					[faculty_details.email],
-					fail_silently=False,
-				)
-			return Response('Success')
-		except ObjectDoesNotExist as notCreated:
-			return Response('Error: Database Integrity Compromised, Contact Support if this persists')
+		check = FacultyListLor.objects.get(lor_id=lor, faculty_id=faculty)
+		if check.application_status== 'AC' or check.application_status== 'AP':
+			FacultyListLor.objects.get(lor=lor, faculty=faculty).delete()
+			send_mail(
+				'Request Withdrawn: Lor Request has been Withdrawn by the student',
+				template.render(context=Context({'faculty': faculty_details, 'student': details})),
+				'ghotden@gmail.com',
+				[faculty_details.email],
+				fail_silently=False,
+			)
+			return Response('Application Withdrawn')
+		else:
+			raise ValidationError({'error_delete': 'You cannot withdraw application at this stage'})
 
 
 # List of Universities in database
